readme

+++ Feral Wiki/Installable software/Restarting - rtorrent - Deluge - Transmission - MySQL/scripts/restart.sh
@@ -1,18 +1,24 @@
 #!/bin/bash
 # restart.sh
-scriptversion="1.0.4"
+scriptversion="1.0.5"
 scriptname="restart"
 # randomessence
 #
 # wget -qO ~/restart http://git.io/5Uw8Gw && bash ~/restart
 #
 ############################
-## Version History Starts ##
+#### Script Notes Start ####
 ############################
 #
-# v1.0.2 templated updated
-# v1.0.3 templated updated
-# v1.0.4 directory exists checks
+# Add notes or warnings here for anyone modifying the scripts
+#
+############################
+##### Script Notes End #####
+############################
+#
+############################
+## Version History Starts ##
+############################
 #
 ############################
 ### Version History Ends ###
@@ -21,6 +27,8 @@ scriptname="restart"
 ############################
 ###### Variable Start ######
 ############################
+#
+updaterenabled="1"
 #
 scripturl="https://raw.github.com/feralhosting/feralfilehosting/master/Feral%20Wiki/Installable%20software/Restarting%20-%20rtorrent%20-%20Deluge%20-%20Transmission%20-%20MySQL/scripts/restart.sh"
 #
@@ -32,26 +40,32 @@ scripturl="https://raw.github.com/feralhosting/feralfilehosting/master/Feral%20W
 #### Self Updater Start ####
 ############################
 #
-[[ ! -d ~/bin ]] && mkdir -p ~/bin
-[[ ! -f ~/bin/"$scriptname" ]] && wget -qO ~/bin/"$scriptname" "$scripturl"
-#
-wget -qO ~/.000"$scriptname" "$scripturl"
-#
-if [[ $(sha256sum ~/.000"$scriptname" | awk '{print $1}') != $(sha256sum ~/bin/"$scriptname" | awk '{print $1}') ]]
+if [[ "$updaterenabled" -eq 1 ]]
 then
-    echo -e "#!/bin/bash\nwget -qO ~/bin/$scriptname $scripturl\ncd && rm -f $scriptname{.sh,}\nbash ~/bin/$scriptname\nexit" > ~/.111"$scriptname"
-    bash ~/.111"$scriptname"
-    exit
-else
-    if [[ -z $(ps x | fgrep "bash $HOME/bin/$scriptname" | grep -v grep | head -n 1 | awk '{print $1}') && $(ps x | fgrep "bash $HOME/bin/$scriptname" | grep -v grep | head -n 1 | awk '{print $1}') -ne "$$" ]]
+    [[ ! -d ~/bin ]] && mkdir -p ~/bin
+    [[ ! -f ~/bin/"$scriptname" ]] && wget -qO ~/bin/"$scriptname" "$scripturl"
+    #
+    wget -qO ~/.000"$scriptname" "$scripturl"
+    #
+    if [[ $(sha256sum ~/.000"$scriptname" | awk '{print $1}') != $(sha256sum ~/bin/"$scriptname" | awk '{print $1}') ]]
     then
-        echo -e "#!/bin/bash\ncd && rm -f $scriptname{.sh,}\nbash ~/bin/$scriptname\nexit" > ~/.222"$scriptname"
-        bash ~/.222"$scriptname"
+        echo -e "#!/bin/bash\nwget -qO ~/bin/$scriptname $scripturl\ncd && rm -f $scriptname{.sh,}\nbash ~/bin/$scriptname\nexit" > ~/.111"$scriptname"
+        bash ~/.111"$scriptname"
         exit
+    else
+        if [[ -z $(ps x | fgrep "bash $HOME/bin/$scriptname" | grep -v grep | head -n 1 | awk '{print $1}') && $(ps x | fgrep "bash $HOME/bin/$scriptname" | grep -v grep | head -n 1 | awk '{print $1}') -ne "$$" ]]
+        then
+            echo -e "#!/bin/bash\ncd && rm -f $scriptname{.sh,}\nbash ~/bin/$scriptname\nexit" > ~/.222"$scriptname"
+            bash ~/.222"$scriptname"
+            exit
+        fi
     fi
+    cd && rm -f .{000,111,222}"$scriptname"
+    chmod -f 700 ~/bin/"$scriptname"
+else
+    echo
+    echo "The Updater has been disabled"
 fi
-cd && rm -f .{000,111,222}"$scriptname"
-chmod -f 700 ~/bin/"$scriptname"
 #
 ############################
 ##### Self Updater End #####
@@ -94,7 +108,7 @@ then
                         then
                             echo -e "\033[31m""Killing all instances of rtorrent""\e[0m"
                             echo
-                            killall -9 -u $(whoami) rtorrent
+                            killall -u $(whoami) rtorrent
                             screen -wipe > /dev/null 2>&1
                             echo "Restarting rtorrent"
                             echo
@@ -123,7 +137,7 @@ then
                         if [[ -d ~/private/deluge ]]
                         then
                             echo -e "\033[31m""Stopping Deluge and Deluge Web""\e[0m"
-                            killall -9 -u $(whoami) deluged deluge-web
+                            killall -u $(whoami) deluged deluge-web
                             echo "Restarting Deluge"
                             deluged
                             echo "Restarting Deluge Web"
@@ -146,7 +160,7 @@ then
                         if [[ -d ~/private/transmission ]]
                         then
                             echo -e "\033[31m""Restarting Transmission""\e[0m"
-                            killall -9 -u $(whoami) transmission-daemon
+                            killall -u $(whoami) transmission-daemon
                             echo "It can take up to 5 minutes for transmission to restart."
                             echo
                             echo "Do this command to see if it is running":
@@ -168,7 +182,7 @@ then
                                 echo -e "\033[31m""Restarting MySQL""\e[0m"
                                 killall -9 -u $(whoami) mysqld mysqld_safe
                                 bash ~/private/mysql/launch.sh > /dev/null 2>&1
-                                echo "Mysql restarted"
+                                echo "Mysql has been restarted"
                                 echo
                                 echo -e "\033[32m""For troubleshooting refer to the FAQ:""\e[0m" "\033[36m""https://www.feralhosting.com/faq/view?question=158""\e[0m"
                                 echo
